pyproject.toml
[project]
name = "kendocenter"
version = "0.2.0"
description = "YKC Kenjin — AI assistant for Yushinkai Kendo Club"
license = "AGPL-3.0-only"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "pydantic-settings>=2.0.0",
    "aiosqlite>=0.20.0",
    "streamlit>=1.40.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["src"]
